Refactor item data classes and update references
- Removed JsonIgnore attributes from various fields in RaceData and RealmData classes.
- Implemented Clone method in RaceData and RealmData classes for deep copying.
- Renamed EquitmentData to EquipmentData across multiple scripts for consistency.
- Changed itemPrice type from int to ulong in ItemPreset and related classes.
- Updated InventoryCenterManager and related services to use EquipmentData.
- Refactored SaveLoadClient services to utilize GameDataCenterManager for item retrieval.
- Cleaned up LoadGameData classes to handle EquipmentData correctly.
- Adjusted UI components to reflect changes in EquipmentData naming.
- Updated ShopDataPopup and BuyItemPopup to handle ulong price type.

+++ Assets/_Project/Scripts/Data/OS/Items/ItemPreset.cs
@@ -16,7 +16,7 @@ public abstract class ItemPreset : ScriptableObject
     public string itemFilePath;
 
     public string itemDescription;
-    public int itemPrice = 100;
+    public ulong itemPrice = 100;
     public int currentstack;
     public bool canStack = false;
     public QualityType qualityType;
